Updating to .NET Generic Host stops using InMemory Database
Following the suggestion of overriding the CreateHost to resolve
any needed services, I got first into an issue of not actually having
called the CreateHost I've overridden. This seems is called only once
the IWebHostBuilder is changed with IHostBuilder in Program.cs
(https://github.com/dotnet-architecture/eShopOnWeb/issues/465)

After this change, the tests aren't using the InMemory database
anymore, but they actually run against the SqlServer, thus
ConfigureServices or ConfigureTestServices is not used as expected.